New deterministic gas implementation (#179)
* New deterministic gas implementation

* Merge master into wojtek/deterministic-gas-wrapper

* Merge master into wojtek/deterministic-gas-wrapper

* Merge remote-tracking branch 'origin/master' into wojtek/deterministic-gas-wrapper

# Conflicts:
#	integration-tests/auth/gas.go
#	integration-tests/testing/helpers.go

* Merge master into wojtek/deterministic-gas-wrapper

* fixes

* Merge master into wojtek/deterministic-gas-wrapper

* Merge remote-tracking branch 'origin/master' into wojtek/deterministic-gas-wrapper

# Conflicts:
#	integration-tests/testing/network.go
#	pkg/config/network.go

* after-merge fixes

* Move deterministic gas config to `pkg/config`

* Remove leftover

* fixes

* Merge master into wojtek/deterministic-gas-wrapper

* Merge master into wojtek/deterministic-gas-wrapper

* fixes

* Merge master into wojtek/deterministic-gas-wrapper

* Merge master into wojtek/deterministic-gas-wrapper

* fixes

* Move deterministic gas types to separate package

+++ pkg/client/client.go
@@ -148,7 +148,7 @@ func (c Client) Encode(signedTx authsigning.Tx) []byte {
 // BroadcastResult contains results of transaction broadcast
 type BroadcastResult struct {
 	TxHash    string
-	GasUsed   int64
+	GasUsed   uint64
 	EventLogs sdk.StringEvents
 }
 
@@ -218,7 +218,7 @@ func (c Client) Broadcast(ctx context.Context, encodedTx []byte) (BroadcastResul
 	}
 	return BroadcastResult{
 		TxHash:    txHash,
-		GasUsed:   resultTx.TxResult.GasUsed,
+		GasUsed:   uint64(resultTx.TxResult.GasUsed),
 		EventLogs: sdk.StringifyEvents(resultTx.TxResult.Events),
 	}, nil
 }
